requirements.txt
Flask>=3.0,<4.0
flask-cors>=4.0.0
pytest>=7.0,<8.0
openai>=1.0.0
azure-cosmos>=4.5.0
azure-storage-blob>=12.17.0
azure-identity>=1.12.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
@jinzcdev/leetcode-mcp-server>=1.2.0
